Collect and upload location data on capture/adding a Tree
This change addresses the issue 438. When adding a tree, in addition to
the data collected when capture a picture of the tree, GPS data stream
obtained by the device is collected and is stored in the device for
later upload for analysis. The duration of this data stream is limited
only during the time the user is presented with the camera view on
adding a Tree.

Tests:
1. Tested the entries of location data entity stored in the Room table
after taking pictures of trees.

./adb -e shell
sqlite3
/data/data/org.greenstand.android.TreeTracker.dev/databases/treetracker.v1.db
select * from location_data;

2. Verified the data uploaded to the s3 after syncing the data from the
device