User items (#678)
* Removed note feature from selfie camera
Localized strings and

* Formatted to remove unnecessary spacing

* Formatted to remove unnecessary spacing

* Made reusable + circle button ISSUE #669 and bug fix ISSUE #668

* Fixed Bug that prevented user from adding creditials(email and phone) to the database.
The bug was as a result of having different userstate instances in the livedata from two composes. Other possible solutions could have been creating a singleton of the data class which is generally a bad practice. This was the reason for creating a new SignUpScreen compose file for both screens so they can always use the same instance of viewmodel and UserState

* Signup flow validation and navigation as requested in #664

* Selfie Screen Aspect ratio #671 FIXed

* User and wallet selection items to update credentials as mails or phone depending on what is available.

* Requested changes from previous pull request https://github.com/Greenstand/treetracker-android/pull/674

* Requested changes from previous pull request https://github.com/Greenstand/treetracker-android/pull/674

* Changed WalletSelectScreen.kt to pass CI/CD test.

Co-authored-by: Jonathan Muller <elforama@gmail.com>